+++ lib/src/format/money.dart
@@ -1,6 +1,9 @@
-import '../constant/money.dart';
-import '../convert/money.dart';
-import '../format/number.dart';
+import 'dart:math';
+
+import 'package:hive_dart/src/constant/money.dart';
+import 'package:hive_dart/src/convert/money.dart';
+import 'package:hive_dart/src/format/number.dart';
+import 'package:hive_dart/src/util/number.dart';
 
 String formatAmount(int value, { unit = '元' }) {
   return formatNumberWithComma(moneyToDisplay(value), decimals: 2) + unit;
@@ -8,4 +11,15 @@ String formatAmount(int value, { unit = '元' }) {
 
 String formatPenny(int value, { unit = '元' }) {
   return formatNumberWithComma(moneyToDisplay(value, unit: MONEY_YUAN_TO_PENNY), decimals: 3) + unit;
+}
+
+String formatAmountShortly(int value, {String unit = '元'}) {
+  return '${
+    shortNumber(
+      moneyToDisplay(value),
+      (num value) {
+        return truncateNumber(value, decimals: 2);
+      }
+    )
+  }$unit';
 }