+++ lib/src/util/number.dart
@@ -1,4 +1,5 @@
 import 'package:decimal/decimal.dart';
+import 'package:hive_dart/src/is/number.dart';
 
 /// 将 `num`（int/double）转换为 `Decimal`，避免精度问题
 Decimal _numToDecimal(num n) {
@@ -53,4 +54,21 @@ String truncateNumber(num value, {int decimals = 0}) {
       : decimalPart.padRight(decimals, '0');
 
   return '$integerPart.$truncatedDecimal';
+}
+
+/// 以较短的方式返回数字
+String shortNumber(num value, String Function(num) formatUnshort) {
+  if (value >= 1000000000000) {
+    final trillion = divideNumber(value, 1000000000000);
+    return '${truncateNumber(trillion, decimals: isInteger(trillion) ? 0 : 1)}万亿';
+  }
+  if (value >= 100000000) {
+    final billion = divideNumber(value, 100000000);
+    return '${truncateNumber(billion, decimals: isInteger(billion) ? 0 : 1)}亿';
+  }
+  if (value >= 10000) {
+    final tenThousand = divideNumber(value, 10000);
+    return '${truncateNumber(tenThousand, decimals: isInteger(tenThousand) ? 0 : 1)}万';
+  }
+  return formatUnshort(value);
 }