+++ test/format_test.dart
@@ -141,6 +141,20 @@ void main() {
       expect(formatAmount(1230), '12.30元');
       expect(formatAmount(1234), '12.34元');
 
+      expect(formatAmountShortly(100000000000000), '1万亿元');
+      expect(formatAmountShortly(110000000000000), '1.1万亿元');
+      expect(formatAmountShortly(111000000000000), '1.1万亿元');
+      expect(formatAmountShortly(100000000000), '10亿元');
+      expect(formatAmountShortly(10000000000), '1亿元');
+      expect(formatAmountShortly(11000000000), '1.1亿元');
+      expect(formatAmountShortly(11100000000), '1.1亿元');
+      expect(formatAmountShortly(10000000), '10万元');
+      expect(formatAmountShortly(1000000), '1万元');
+      expect(formatAmountShortly(1100000), '1.1万元');
+      expect(formatAmountShortly(1110000), '1.1万元');
+      expect(formatAmountShortly(1000), '10.00元');
+      expect(formatAmountShortly(1000, unit: ''), '10.00');
+
     });
 
     test('formatPenny', () {